% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_db.R
\name{as_gene_sets}
\alias{as_gene_sets}
\title{Extract Gene Sets from a Pathway Database}
\usage{
as_gene_sets(pathway_db)
}
\arguments{
\item{pathway_db}{An \code{EnrichKit_pathway_db} object or a named list.}
}
\value{
A named list of gene sets.
}
\description{
Extract Gene Sets from a Pathway Database
}
