% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_symbol.R
\name{check_symbol}
\alias{check_symbol}
\title{Suspicious gene symbol checking}
\usage{
check_symbol(notsure)
}
\arguments{
\item{notsure}{Single or multiple suspicious gene symbols to be checked. multiple elements should be in an vector.}
}
\value{
Single or multiple "suggested" gene symbols. If the original symbols are approved or invalid, return themselves as so, otherwise return the HGNC approved suggested symbols.
  Refer to \code{\link{checkGeneSymbols}} from package \code{\link{HGNChelper}}
}
\description{
Check whether the current gene symbols are the approved HGNC suggested gene symbols, associate the latter if not
}
\examples{
check_symbol(c("TNK2","ACR"))

}
