% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNformat.R
\name{ConvertNformat}
\alias{ConvertNformat}
\title{A funcition to convert Ensembl ID to Entrez ID.}
\usage{
ConvertNformat(
  bg_gene,
  TestingSubsetNames,
  TestingModAssign,
  keyword = "Ensembl2Entrez_Convert"
)
}
\arguments{
\item{bg_gene}{A vector of genes with Ensembl id as format}

\item{TestingSubsetNames}{A vector that has the names of each module, indicates the number of test will be run.}

\item{TestingModAssign}{A vector of module assignment of each gene id}

\item{keyword}{A keyword that will be included in all the outputs, helps user to recognize.}
}
\value{
Nothing, but will output a .RData file will all the enrichment records.
}
\description{
A funcition to convert Ensembl ID to Entrez ID.
}
