% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNformatID.R
\name{convertNformatID}
\alias{convertNformatID}
\title{Match different identifiers to gene sets provided}
\usage{
convertNformatID(GeneSetNames, SigGene_list, TotalGene_list, IDtype = "ens")
}
\arguments{
\item{GeneSetNames}{Sigle or multiple gene set names, put into a vector if multiple. e.g. c('Lactation1','Lactation2','Lactation3')}

\item{SigGene_list}{An list object; where each element contains signifcant genes. list dimension must match \code{TotalGene_list}}

\item{TotalGene_list}{An list object; where each element contains total genes. list dimension must match \code{SigGene_list}}

\item{IDtype}{Gene Identifier type, choose from \code{c("ens","entrez","symbol")}}
}
\value{
An list object with same dimension as SigGene_list and Totalgene_list, where each element has comprehensive id information.
}
\description{
Match different identifiers to gene sets provided
}
\examples{

}
