% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_Enrich_Plot.R
\name{Go_Enrich_Plot}
\alias{Go_Enrich_Plot}
\title{A function systematically performs hypergeomatric-based over-representation analysis in Gene Ontology database.}
\usage{
Go_Enrich_Plot(
  total_genes_all,
  sig_genes_all,
  TestingSubsetNames,
  GOthres = 0.05,
  biomart = "ensembl",
  dataset = "btaurus_gene_ensembl",
  attributes = c("ensembl_gene_id", "go_id", "name_1006"),
  keyword = "GO_Enrichment_thres_point1_5sets"
)
}
\arguments{
\item{total_genes_all}{A list of single or multiple elements, contains all back ground genes for each test, should have the same length with sig_genes_all.}

\item{sig_genes_all}{A list of single or multiple elements, contains all significant genes for each test, should have the same length with total_genes_all.}

\item{TestingSubsetNames}{A vector that has the names of each module, indicates the number of test will be run, should have the same length with total_genes_all/sig_genes_all.}

\item{GOthres}{P-value threshold to determain significance of each records.}

\item{biomart}{Name of annotation database. for now, only have "ensembl"}

\item{dataset}{Name of annotation database. e.g. "btaurus_gene_ensembl"}

\item{attributes}{A vector contains all attributes to be retrieved from the database, e.g. \code{c("ensembl_gene_id","go_id","name_1006")}}

\item{keyword}{A keyword that will be included in all the outputs, helps user to recognize.}
}
\value{
Nothing, but will output a .RData file will all the enrichment records.
}
\description{
A function systematically performs hypergeomatric-based over-representation analysis in Gene Ontology database.
}
