% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGEnrich.R
\name{HyperGEnrich}
\alias{HyperGEnrich}
\title{Over-representation test process}
\usage{
HyperGEnrich(
  GeneSet = sigGene_All,
  Database = "",
  minOverlap = 4,
  pvalue_thres = 0.05,
  adj_pvalue_thres = 1,
  padj_method = "BH"
)
}
\arguments{
\item{GeneSet}{A list object; Ideal input is object resulted from \codes{\links{convertNformatID}}
Each element is a gene set to be tested, should have different identifiers in columns, 1st column - ensembl id; 2nd - entrezid,
Also should have a column "Sig" indicating significant or not, 1 - significant; 0 insignificant}

\item{Database}{Indicate one of the dataset to be used; choose from c("go","kegg","interpro","mesh","msig","reactome")}

\item{minOverlap}{Minimum overlap number of total genes and genes of a target pathway, default value is \code{4}}

\item{pvalue_thres}{Pvalue of the Fisher's exact test. Refer to \code{\link{fisher.test}}}

\item{adj_pvalue_thres}{Adjusted pvalues; choose methods from c("bonferroni","hochberg","BH"). Refer to \code{\link{p.adjust}}}

\item{padj_method}{}
}
\value{

}
\description{
Over-representation test process
}
\examples{

}
