% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpro_DB_Update.R
\name{Interpro_DB_Update}
\alias{Interpro_DB_Update}
\title{Update/download new database - Interpro.
New database will be gathered and stored locally.}
\usage{
Interpro_DB_Update(
  biomart = "ensembl",
  dataset = "btaurus_gene_ensembl",
  attributes = c("ensembl_gene_id", "interpro", "interpro_description"),
  keyword = "Interpro_DB",
  DB_location = "."
)
}
\arguments{
\item{biomart}{BioMart database name you want to connect to. Default \code{biomart="ensembl"}}

\item{dataset}{Mart object created with the useMart function. Default \code{dataset="btaurus_gene_ensembl"}}

\item{attributes}{Attributes you want to retrieve. Refer to \code{listAttributes}.}

\item{keyword}{Keyword used for naming. Default \code{keyword = "Interpro_DB"}}

\item{DB_location}{Store path of the new database. Default is current working directory.}
}
\value{
NA. New dataset will be packed and stored in .RData format.
}
\description{
Update/download new database - Interpro.
New database will be gathered and stored locally.
}
\examples{
Interpro_DB_Update()

}
