% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterPro_Enrich.R
\name{InterPro_Enrich}
\alias{InterPro_Enrich}
\title{A function systematically performs hypergeomatric-based over-representation analysis in Interpro Database.}
\usage{
InterPro_Enrich(
  total_genes_all,
  sig_genes_all,
  TestingSubsetNames,
  IPthres = 0.05,
  biomart = "ensembl",
  dataset = "btaurus_gene_ensembl",
  Identifier = "external_gene_name",
  attributes = c("ensembl_gene_id", "external_gene_name", "interpro",
    "interpro_description"),
  keyword = "Interpro_Enrichment"
)
}
\arguments{
\item{total_genes_all}{A list of single or multiple elements, contains all back ground genes for each test, should have the same length with sig_genes_all.}

\item{sig_genes_all}{A list of single or multiple elements, contains all significant genes for each test, should have the same length with total_genes_all.}

\item{TestingSubsetNames}{A vector that has the names of each module, indicates the number of test will be run, should have the same length with total_genes_all/sig_genes_all.}

\item{IPthres}{P-value threshold to determain significance of each records.}

\item{biomart}{Name of annotation database. for now, only have "ensembl".}

\item{dataset}{Name of annotation database. e.g. "btaurus_gene_ensembl".}

\item{Identifier}{Whether "enternal_gene_name" or "ensembl_gene_id" will be used as the identifier in the imput dataset.}

\item{attributes}{A vector contains all attributes to be retrieved from the database, e.g. \code{c("ensembl_gene_id","go_id","name_1006")}}

\item{keyword}{A keyword that will be included in all the outputs, helps user to recognize.}
}
\value{
Nothing, but will output a .RData file will all the enrichment records.
}
\description{
A function systematically performs hypergeomatric-based over-representation analysis in Interpro Database.
}
