% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_DB_Update.R
\name{Kegg_DB_Update}
\alias{Kegg_DB_Update}
\title{Update/download new database - KEGG.
New database will be gathered and stored locally.}
\usage{
Kegg_DB_Update(
  species = "bta",
  id.type = "kegg",
  keyword = "KEGG_DB",
  DB_location = "."
)
}
\arguments{
\item{species}{Default \code{species = "bta"}}

\item{id.type}{Default \code{id.type = "kegg"}}

\item{keyword}{Keyword used for naming. Default \code{keyword = "KEGG_DB"}}

\item{DB_location}{Store path of the new database. Default is current working directory.}
}
\value{
NA. New dataset will be packed and stored in .RData format.
}
\description{
Update/download new database - KEGG.
New database will be gathered and stored locally.
}
\examples{
Kegg_DB_Update()

}
