% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kegg_Enrich_Plot.R
\name{Kegg_Enrich_Plot}
\alias{Kegg_Enrich_Plot}
\title{A function systematically performs hypergeomatric-based over-representation analysis in KEGG Database.}
\usage{
Kegg_Enrich_Plot(
  sig_genes_all,
  total_genes_all,
  TestingSubsetNames,
  KEGGthres = 0.05,
  species = "bta",
  id.type = "kegg",
  Sig_list_out = Sig_list_out,
  keyword
)
}
\arguments{
\item{sig_genes_all}{A list of single or multiple elements, contains all back ground genes for each test, should have the same length with sig_genes_all.}

\item{total_genes_all}{A list of single or multiple elements, contains all significant genes for each test, should have the same length with total_genes_all.}

\item{TestingSubsetNames}{A vector that has the names of each module, indicates the number of test will be run, should have the same length with total_genes_all/sig_genes_all.}

\item{KEGGthres}{P-value threshold to determain significance of each records.}

\item{species}{species = "bta"}

\item{id.type}{id.type = "kegg"}

\item{Sig_list_out}{A list contains all identifier conversion information, used to print out the significant genes in each enrichment records, in order to have the same type of identifier with imput.}

\item{keyword}{A keyword that will be included in all the outputs, helps user to recognize.}
}
\value{
Nothing, but will output a .RData file will all the enrichment records.
}
\description{
A function systematically performs hypergeomatric-based over-representation analysis in KEGG Database.
}
