% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kfirst_gosets_source}
\alias{kfirst_gosets_source}
\title{Kids First Pathway Source Table}
\format{
A data frame with columns:
\describe{
\item{pathway}{Pathway/gene-set name.}
\item{source}{Source database label.}
\item{n_genes}{Number of retained genes after Kids First filtering.}
}
}
\usage{
kfirst_gosets_source
}
\description{
Source/provenance table for \link{kfirst_gosets_all}.
}
\keyword{datasets}
