% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.R
\name{make_pathway_db_from_gmt}
\alias{make_pathway_db_from_gmt}
\title{Build a Pathway Database from GMT}
\usage{
make_pathway_db_from_gmt(
  file,
  database = tools::file_path_sans_ext(basename(file)),
  source = "GMT",
  version = NA_character_,
  universe = NULL,
  min_size = 1,
  max_size = Inf
)
}
\arguments{
\item{file}{GMT file path.}

\item{database}{Database label to assign to all GMT pathways.}

\item{source}{Source label.}

\item{version}{Optional version label.}

\item{universe}{Optional analysis universe.}

\item{min_size}{Minimum retained set size.}

\item{max_size}{Maximum retained set size.}
}
\value{
An \code{EnrichKit_pathway_db} object.
}
\description{
Build a Pathway Database from GMT
}
