% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeSH_DB_Update.R
\name{MeSH_DB_Update}
\alias{MeSH_DB_Update}
\title{Update/download new database - Mesh.
New database will be gathered and stored locally.}
\usage{
MeSH_DB_Update(keyword = "MeSH_DB", DB_location = ".")
}
\arguments{
\item{keyword}{Keyword used for naming. Default \code{keyword = "MeSH_DB"}}

\item{DB_location}{Store path of the new database. Default is current working directory.}
}
\value{
NA. New dataset will be packed and stored in .RData format.
}
\description{
Update/download new database - Mesh.
New database will be gathered and stored locally.
}
\examples{
MeSH_DB_Update()

}
