% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeSH_Enrich.R
\name{MeSH_Enrich}
\alias{MeSH_Enrich}
\title{A function systematically performs hypergeomatric-based over-representation analysis in MeSH Database.}
\usage{
MeSH_Enrich(
  total_genes_all,
  sig_genes_all,
  TestingSubsetNames,
  Meshthres = 0.05,
  Sig_list_out,
  MeshCate = c("D", "G"),
  dataset = "MeSH.Bta.eg.db",
  keyword = "MESH_Enrichment"
)
}
\arguments{
\item{total_genes_all}{A list of single or multiple elements, contains all back ground genes for each test, should have the same length with sig_genes_all.}

\item{sig_genes_all}{A list of single or multiple elements, contains all significant genes for each test, should have the same length with total_genes_all.}

\item{TestingSubsetNames}{A vector that has the names of each module, indicates the number of test will be run, should have the same length with total_genes_all/sig_genes_all.}

\item{Meshthres}{P-value threshold to determain significance of each records.}

\item{Sig_list_out}{A list contains all identifier conversion information, used to print out the significant genes in each enrichment records, in order to have the same type of identifier with imput.}

\item{MeshCate}{Mesh category. for now, only "D" and "G"}

\item{dataset}{Species specific. e.g. Bos Taurus - "MeSH.Bta.eg.db"}

\item{keyword}{A keyword that will be included in all the outputs, helps user to recognize.}
}
\value{
Nothing, but will output a .RData file will all the enrichment records.
}
\description{
A function systematically performs hypergeomatric-based over-representation analysis in MeSH Database.
}
