% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msig_DB_Update.R
\name{Msig_DB_Update}
\alias{Msig_DB_Update}
\title{Update/download new database - Msig.
New database will be gathered and stored locally.
Note this function could take quite a while...}
\usage{
Msig_DB_Update(keyword = "Msig_DB", DB_location = ".")
}
\arguments{
\item{keyword}{Keyword used for naming. Default \code{keyword = "Msig_DB"}}

\item{DB_location}{Store path of the new database. Default is current working directory.}
}
\value{
NA. New dataset will be packed and stored in .RData format.
}
\description{
Update/download new database - Msig.
New database will be gathered and stored locally.
Note this function could take quite a while...
}
\examples{
Msig_DB_Update()

}
