% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse_Results.R
\name{Parse_Results}
\alias{Parse_Results}
\title{Parse results from over-representation test, create a new data frame contains all significant hits}
\usage{
Parse_Results(Results_List, keyword = "Which D.B")
}
\arguments{
\item{Results_List}{A Returning list of single or multiple element.}

\item{keyword}{A note just to remind user of the database name, will be included in the prompts.}
}
\value{
A data frame contains all enriched records under pre-defined significance value
}
\description{
Parse results from over-representation test, create a new data frame contains all significant hits
}
