% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_db.R
\name{pathway_gene_counts}
\alias{pathway_gene_counts}
\title{Count Pathway Members in a Background}
\usage{
pathway_gene_counts(pathway_db, background = NULL)
}
\arguments{
\item{pathway_db}{An \code{EnrichKit_pathway_db} object or named list.}

\item{background}{Optional character vector of valid analysis features.}
}
\value{
A data frame with pathway gene counts.
}
\description{
Count Pathway Members in a Background
}
