% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_db.R
\name{pathway_metadata}
\alias{pathway_metadata}
\title{Extract Pathway Metadata}
\usage{
pathway_metadata(pathway_db)
}
\arguments{
\item{pathway_db}{An \code{EnrichKit_pathway_db} object.}
}
\value{
A data frame with pathway metadata.
}
\description{
Extract Pathway Metadata
}
