% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{pathway_similarity}
\alias{pathway_similarity}
\title{Pairwise Pathway Similarity}
\usage{
pathway_similarity(gene_sets)
}
\arguments{
\item{gene_sets}{Named list of gene sets or an \code{EnrichKit_pathway_db} object.}
}
\value{
A data frame of pairwise pathway similarities.
}
\description{
Calculates pairwise Jaccard similarity and gene separation between pathway
gene sets.
}
