% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_DB_Update.R
\name{Reactome_DB_Update}
\alias{Reactome_DB_Update}
\title{Update/download new database - Reactome.
New database will be gathered and stored locally.}
\usage{
Reactome_DB_Update(
  websource = "https://reactome.org/download/current/NCBI2Reactome_All_Levels.txt",
  Species = "Bos taurus",
  keyword = "Reactome_DB",
  DB_location = "."
)
}
\arguments{
\item{websource}{\code{websource = "https://reactome.org/download/current/NCBI2Reactome_All_Levels.txt"}}

\item{Species}{Default \code{Species = "Bos taurus"}}

\item{keyword}{Keyword used for naming. Default \code{keyword = "Reactome_DB"}}

\item{DB_location}{Store path of the new database. Default is current working directory.}
}
\value{
NA. New dataset will be packed and stored in .RData format.
}
\description{
Update/download new database - Reactome.
New database will be gathered and stored locally.
}
\examples{
Reactome_DB_Update()

}
