% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceDim_GO_Plot.R
\name{ReduceDim_GO_Plot}
\alias{ReduceDim_GO_Plot}
\title{A function to retrieve all the pair-wise semantic similarity of each enriched GO record and visualize the relationship in the format of correlation plots.}
\usage{
ReduceDim_GO_Plot(
  Enrich_Out,
  GOthres = 0.001,
  label_sizeCC = 0.4,
  label_sizeBP = 0.4,
  label_sizeMF = 0.4,
  Database = "org.Bt.eg.db",
  measure = "Jiang",
  combine = NULL,
  Dataset_Name
)
}
\arguments{
\item{Enrich_Out}{A vector of significantly enriched GO terms.}

\item{GOthres}{The threshold used to plot.}

\item{label_sizeCC}{The font size of category CC}

\item{label_sizeBP}{The font size of category BP}

\item{label_sizeMF}{The font size of category MF}

\item{Database}{The database used to retrieve the semantic similarity measurement}

\item{measure}{Measurement of semantic similarity, consistant with \code{GOSemSim} package.}

\item{combine}{NULL}

\item{Dataset_Name}{A keyword that will be included in the output names, helps users to recognize.}
}
\value{
Nothing, but will generate PDF plots and the corresponding correlation matrix.
}
\description{
A function to retrieve all the pair-wise semantic similarity of each enriched GO record and visualize the relationship in the format of correlation plots.
}
