% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{signed_log10_p}
\alias{signed_log10_p}
\title{Signed -log10 P-value}
\usage{
signed_log10_p(p, direction)
}
\arguments{
\item{p}{Numeric p-values.}

\item{direction}{Numeric direction/effect sign.}
}
\value{
Signed \code{-log10(p)}.
}
\description{
Signed -log10 P-value
}
