% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tibble.R
\name{split_tibble}
\alias{split_tibble}
\title{Parse/split a tibble into a list according to one column.
Ideally one column has categorical variable thus can be eatracted, each level would become a new element name of the list.}
\usage{
split_tibble(tibble, column, keep)
}
\arguments{
\item{tibble}{a tibble object that has at least two columns.}

\item{column}{String. The column used as index, preferably contains categorical variable.}

\item{keep}{String. The single or multiple column names to keep.}
}
\value{
A list. The selected column with preferably differetn levels would be used as index and the rest or selected columns would be kept.
The tibble would become a list.
}
\description{
Parse/split a tibble into a list according to one column.
Ideally one column has categorical variable thus can be eatracted, each level would become a new element name of the list.
}
\examples{
df = tibble(A = c(1:5),B = c(rep('M',3),rep('T',2)))
split_tibble(df,'B','A')


}
