% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumer.R
\name{write_gmt}
\alias{write_gmt}
\title{Write a Named Gene-Set List as GMT}
\usage{
write_gmt(gene_sets, file)
}
\arguments{
\item{gene_sets}{Named list of character vectors.}

\item{file}{Output GMT path.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Write a Named Gene-Set List as GMT
}
