% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPre_C.R
\name{DataPre_C}
\alias{DataPre_C}
\title{RNA-seq data preparation for correlation network analysis - With pc correction}
\usage{
DataPre_C(
  networkData,
  cousin = 0.4,
  n1,
  n2,
  perct,
  correct = T,
  thres_rmzero = 5,
  count_rmzero
)
}
\arguments{
\item{networkData}{A dataframe contains expression data, samples in rows and genes in columns.}

\item{cousin}{Percentage of reads that need to be removed. Only a few genes have many reads thus we remove them.}

\item{n1}{Number of samples in referece group.}

\item{n2}{Number of samples in test group.}

\item{perct}{Percentage of genes need to be removed based on overall expression variance. Only genes with high variance will be included in network analysis.}

\item{correct}{Whether pc-based correction should be used or not.}

\item{thres_rmzero}{Lowest count considered as not expressed, default 5 reads per sample.}

\item{count_rmzero}{threshold of individuals to remove a gene, usually 1/2 sample size.}
}
\value{
A dataframe with corrected measurement of expression.
}
\description{
RNA-seq data preparation for correlation network analysis - With pc correction
}
