% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kfirst_gene_universe}
\alias{kfirst_gene_universe}
\title{Kids First Interrogated Gene Universe}
\format{
Character vector of gene symbols.
}
\usage{
kfirst_gene_universe
}
\description{
Gene universe used to filter \link{kfirst_gosets_all}.
}
\keyword{datasets}
